requirements.txt
Flask==3.1.3
pymongo==4.16.0
python-dotenv==1.2.2
pytz==2026.1
requests==2.32.5
